deploy/requirements.txt
ansible==1.9.1
boto==2.32.1
filechunkio==1.5
gnureadline==6.3.3
ipdb==0.8
ipython==2.1.0
python-decouple==2.2
wsgiref==0.1.2
